% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AF1.R
\name{AF1}
\alias{AF1}
\title{Area do cladodio}
\usage{
AF1(Matriz)
}
\arguments{
\item{Matriz}{:Matriz com o conjunto de dados. Deve-se ter organizado nas colunas respectivamente o valor do comprimento (cm) e largura (cm)do cladodio}
}
\value{
A funcao retorna a producao (Kg) por planta estimada por regressao multipla.
}
\description{
Esta funcao estima a area do cladodio da palma forrageira recorrendo a dados morfologicos e regressao.
}
\examples{
data(Dados2)
Cor=Dados2
AnaliseTrilha(Dados2,6)

}
\references{
Referencias artigo agriambi.
}
