% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PalmaForrageira.R
\name{PalmaForrageira}
\alias{PalmaForrageira}
\title{PalmaForrageira (interface grafica)}
\usage{
PalmaForrageira()
}
\value{
A funcao retorna a producao (Kg) por planta estimada por redes neurais artificiais.
}
\description{
Esta funcao ....
}
\examples{
data(Dados2)
Cor=Dados2
AnaliseTrilha(Dados2,6)

}
\references{
Referencias artigo agriambi.
}
