% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Produtividade1.R
\name{Produtividade1}
\alias{Produtividade1}
\title{Produtividade (interface grafica)}
\usage{
Produtividade1(Matriz)
}
\arguments{
\item{Matriz}{:Matriz com o conjunto de dados. Deve-se ter organizado nas colunas respectivamente o valor medio na planta do comprimento (cm), largura (cm), espessura (mm) das raques, altura da planta (m), numero de raques (un) e area total do cladodio (m²)}
}
\value{
A funcao retorna a producao (Kg) por planta estimada por redes neurais artificiais.
}
\description{
Esta funcao estima a produtividade da palma forrageira recorrendo a dados morfologicos e redes neurais artificiais.
}
\examples{
data(Dados2)
Cor=Dados2
AnaliseTrilha(Dados2,6)

}
\references{
Referencias artigo agriambi.
}
